fix(userstories): add filter to inprove performance for dashboard queries (#tg-4600)
